Updated README.md to correct 2 if statements
Fixed two `if` statements that were supposed to be testing for the existence of `~tmp` and `~tmp/ttf`. The statements weren't working properly because both had wrapped the paths in `"` double quotes, which resulted in the `~` not expanding to `$HOME`. Removing the double quotes fixed it.

Also changed the Hack font package from zip to tar.gz file format. Removed `unzip` from the yum install list.